+++ .vim/pack/basic/start/vim-clap/crates/icon/src/lib.rs
@@ -3,11 +3,18 @@
 // pub use constants::*;
 include!(concat!(env!("OUT_DIR"), "/constants.rs"));
 
-use std::{fmt::Display, path::Path};
+use std::fmt::Display;
+use std::path::Path;
 
-pub const DEFAULT_ICON: char = '';
-pub const FOLDER_ICON: char = '';
-pub const DEFAULT_FILER_ICON: char = '';
+/// The type used to represent icons.
+///
+/// This could be changed into different type later,
+/// so functions take and return this type, not `char` or `&str` directly.
+pub type IconType = char;
+
+pub const DEFAULT_ICON: IconType = '';
+pub const FOLDER_ICON: IconType = '';
+pub const DEFAULT_FILER_ICON: IconType = '';
 
 // Each added icon length is 4 bytes.
 pub const ICON_LEN: usize = 4;
@@ -30,6 +37,10 @@ impl Icon {
             Self::Null => None,
             Self::Enabled(icon_kind) => Some(icon_kind),
         }
+    }
+
+    pub fn enabled(&self) -> bool {
+        matches!(self, Self::Enabled(_))
     }
 }
 
@@ -82,6 +93,13 @@ impl<T: AsRef<str>> From<T> for IconKind {
 
 impl IconKind {
     /// Returns a `String` of raw str with icon added.
+    pub fn add_icon_to_text<S: AsRef<str>>(&self, text: S) -> String {
+        let text = text.as_ref();
+        let icon = self.icon(text);
+        format!("{icon} {text}")
+    }
+
+    /// Returns a `String` of raw str with icon added.
     pub fn paint<S: AsRef<str>>(&self, raw_str: S) -> String {
         let fmt = |s| format!("{} {}", s, raw_str.as_ref());
         match *self {
@@ -104,12 +122,6 @@ impl IconKind {
         }
     }
 }
-
-/// The type used to represent icons.
-///
-/// This could be changed into different type later,
-/// so functions take and return this type, not `char` or `&str` directly.
-type IconType = char;
 
 /// Return appropriate icon for the path. If no icon matched, return the specified default one.
 ///