+++ .vim/pack/basic/start/vim-clap/crates/maple_cli/src/command/blines.rs
@@ -1,7 +1,12 @@
+use std::io::BufRead;
+use std::sync::atomic::{AtomicUsize, Ordering};
+
 use anyhow::Result;
 use clap::Parser;
 
 use filter::Source;
+use matcher::Bonus;
+use types::SourceItem;
 
 use crate::app::Params;
 use crate::paths::AbsPathBuf;
@@ -21,16 +26,41 @@ pub struct Blines {
 impl Blines {
     /// Looks for matches of `query` in lines of the current vim buffer.
     pub fn run(&self, params: Params) -> Result<()> {
+        let source_file = std::fs::File::open(&self.input)?;
+
+        let index = AtomicUsize::new(0);
+        let blines_item_stream = || {
+            std::io::BufReader::new(source_file)
+                .lines()
+                .filter_map(|x| {
+                    x.ok().and_then(|line: String| {
+                        let index = index.fetch_add(1, Ordering::SeqCst);
+                        if line.trim().is_empty() {
+                            None
+                        } else {
+                            let item = SourceItem::from(format!("{index} {line}"));
+                            Some(item)
+                        }
+                    })
+                })
+        };
+
+        let filter_context = if let Some(extension) = self
+            .input
+            .extension()
+            .and_then(|s| s.to_str().map(|s| s.to_string()))
+        {
+            params
+                .into_filter_context()
+                .bonuses(vec![Bonus::Language(extension.into())])
+        } else {
+            params.into_filter_context()
+        };
+
         filter::dyn_run(
             &self.query,
-            Source::List(
-                std::fs::read_to_string(&self.input)?
-                    .lines()
-                    .enumerate()
-                    .map(|(idx, item)| format!("{} {}", idx + 1, item))
-                    .map(Into::into),
-            ),
-            params.into_filter_context(),
+            filter_context,
+            Source::List(blines_item_stream()),
         )
     }
 }