+++ .vim/pack/basic/start/vim-clap/crates/maple_cli/src/command/filter.rs
@@ -3,8 +3,9 @@ use std::path::PathBuf;
 
 use anyhow::Result;
 use clap::Parser;
+use subprocess::Exec;
 
-use filter::{subprocess::Exec, FilterContext, Source};
+use filter::{FilterContext, Source};
 use matcher::{Bonus, FuzzyAlgorithm, MatchScope, Matcher};
 use types::SourceItem;
 
@@ -118,8 +119,8 @@ impl Filter {
         } else {
             filter::dyn_run::<std::iter::Empty<_>>(
                 &self.query,
-                self.generate_source(),
                 FilterContext::new(icon, number, winwidth, matcher),
+                self.generate_source(),
             )?;
         }
         Ok(())