+++ .vim/pack/basic/start/vim-clap/crates/maple_cli/src/command/grep/forerunner.rs
@@ -1,5 +1,14 @@
-use super::*;
-use crate::process::light::CommandEnv;
+use std::path::PathBuf;
+
+use anyhow::Result;
+use clap::Parser;
+
+use utility::is_git_repo;
+
+use super::{rg_base_command, rg_command};
+use crate::app::Params;
+use crate::process::light::{CommandEnv, LightCommand};
+use crate::utils::{send_response_from_cache, SendResponse};
 
 #[derive(Parser, Debug, Clone)]
 pub struct RipGrepForerunner {
@@ -10,6 +19,12 @@ pub struct RipGrepForerunner {
     /// Specify the threshold for writing the output of command to a tempfile.
     #[clap(long = "output-threshold", default_value = "30000")]
     output_threshold: usize,
+
+    /// Run without checking if cwd is a git repo.
+    ///
+    /// By default this command only works when cwd is a git repo.
+    #[clap(long)]
+    force_run: bool,
 }
 
 impl RipGrepForerunner {
@@ -40,7 +55,7 @@ impl RipGrepForerunner {
     ) -> Result<()> {
         if !no_cache {
             if let Some(ref dir) = self.cmd_dir {
-                let base_cmd = BaseCommand::new(RG_EXEC_CMD.into(), dir.clone());
+                let base_cmd = rg_base_command(dir);
                 if let Some((total, cache)) = base_cmd.cache_info() {
                     if total > 100000 {
                         send_response_from_cache(&cache, total as usize, SendResponse::Json, icon);
@@ -50,38 +65,55 @@ impl RipGrepForerunner {
             }
         }
 
-        if self.should_skip() {
+        if !self.force_run && self.should_skip() {
             return Ok(());
         }
 
-        let mut std_cmd = StdCommand::new(RG_ARGS[0]);
-        // Do not use --vimgrep here.
-        std_cmd.args(&RG_ARGS[1..]);
+        let dir = match self.cmd_dir {
+            Some(ref dir) => dir.clone(),
+            None => std::env::current_dir()?,
+        };
 
-        if let Some(ref dir) = self.cmd_dir {
-            std_cmd.current_dir(dir);
-        }
-
-        let mut cmd = std_cmd.into_inner();
+        let mut cmd = rg_command(&dir);
 
         let mut light_cmd = LightCommand::new(
             &mut cmd,
-            CommandEnv::new(
-                self.cmd_dir.clone(),
-                number,
-                icon,
-                Some(self.output_threshold),
-            ),
+            CommandEnv::new(Some(dir.clone()), number, icon, Some(self.output_threshold)),
         );
 
-        let cwd = match self.cmd_dir {
-            Some(d) => d,
-            None => std::env::current_dir()?,
-        };
-        let base_cmd = BaseCommand::new(RG_EXEC_CMD.into(), cwd);
-
-        light_cmd.execute(base_cmd)?.print();
+        light_cmd.execute(rg_base_command(dir))?.print();
 
         Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::app::Params;
+
+    // TODO: fix and enable the test in CI
+    #[test]
+    #[ignore]
+    fn ripgrep_forerunner_command_works() {
+        let params = Params::parse_from(&["--no-cache", "--icon=Grep"]);
+
+        let ripgrep_forerunner = RipGrepForerunner::parse_from(&[
+            "",
+            "--cmd-dir",
+            std::env::current_dir()
+                .unwrap()
+                .into_os_string()
+                .as_os_str()
+                .to_str()
+                .unwrap(),
+            "--output-threshold",
+            "100000",
+            "--force-run",
+        ]);
+
+        ripgrep_forerunner
+            .run(params)
+            .expect("Failed to run command `ripgrep-forerunner`");
     }
 }