+++ .vim/pack/basic/start/vim-clap/crates/maple_cli/src/find_usages/search_engine/ctags/mod.rs
@@ -4,26 +4,29 @@ use std::hash::Hash;
 use std::path::{Path, PathBuf};
 
 use anyhow::Result;
-use filter::subprocess::{Exec, Redirection};
 use itertools::Itertools;
 use rayon::prelude::*;
+use subprocess::{Exec, Redirection};
 
 use super::{QueryType, Symbol};
 use crate::find_usages::AddressableUsage;
-use crate::tools::ctags::TagsConfig;
+use crate::tools::ctags::TagsGenerator;
 use crate::utils::ExactOrInverseTerms;
 
 /// `readtags` powered searcher.
 #[derive(Debug, Clone, Eq, PartialEq, Hash)]
 pub struct CtagsSearcher<'a, P> {
-    config: TagsConfig<'a, P>,
     tags_path: PathBuf,
+    tags_generator: TagsGenerator<'a, P>,
 }
 
 impl<'a, P: AsRef<Path> + Hash> CtagsSearcher<'a, P> {
-    pub fn new(config: TagsConfig<'a, P>) -> Self {
-        let tags_path = config.tags_path();
-        Self { config, tags_path }
+    pub fn new(tags_generator: TagsGenerator<'a, P>) -> Self {
+        let tags_path = tags_generator.tags_path();
+        Self {
+            tags_path,
+            tags_generator,
+        }
     }
 
     /// Returns `true` if the tags file already exists.
@@ -32,8 +35,8 @@ impl<'a, P: AsRef<Path> + Hash> CtagsSearcher<'a, P> {
     }
 
     /// Generate the `tags` file.
-    pub fn generate_tags(&self) -> Result<()> {
-        self.config.generate_tags()
+    pub fn generate_tags(&self) -> std::io::Result<()> {
+        self.tags_generator.generate_tags()
     }
 
     pub fn search_usages(
@@ -64,7 +67,7 @@ impl<'a, P: AsRef<Path> + Hash> CtagsSearcher<'a, P> {
     fn build_exec(&self, query: &str, query_type: QueryType) -> Exec {
         // https://docs.ctags.io/en/latest/man/readtags.1.html#examples
         let cmd = Exec::cmd("readtags")
-            .stderr(Redirection::Merge)
+            .stderr(Redirection::None) // Ignore the line: ctags: warning...
             .arg("--tag-file")
             .arg(&self.tags_path)
             .arg("-E")