+++ .vim/pack/basic/start/vim-clap/crates/maple_cli/src/process/tokio.rs
@@ -2,13 +2,12 @@
 
 use std::path::Path;
 
-use anyhow::Result;
 use tokio::process::Command;
 
 /// Builds `Command` from a cmd string which can use pipe.
 ///
 /// This can work with the piped command, e.g., `git ls-files | uniq`.
-pub fn build_command(shell_cmd: impl AsRef<str>) -> Command {
+pub fn shell_command(shell_cmd: impl AsRef<str>) -> Command {
     if cfg!(target_os = "windows") {
         let mut cmd = Command::new("cmd");
         cmd.args(&["/C", shell_cmd.as_ref()]);
@@ -24,19 +23,19 @@ pub fn build_command(shell_cmd: impl AsRef<str>) -> Command {
 #[derive(Debug)]
 pub struct TokioCommand(Command);
 
-impl<T: AsRef<str>> From<T> for TokioCommand {
-    fn from(cmd: T) -> Self {
-        Self(build_command(cmd))
+impl From<Command> for TokioCommand {
+    fn from(cmd: Command) -> Self {
+        Self(cmd)
     }
 }
 
 impl TokioCommand {
     /// Constructs a new instance of [`TokioCommand`].
-    pub fn new(cmd: impl AsRef<str>) -> Self {
-        cmd.into()
+    pub fn new(shell_cmd: impl AsRef<str>) -> Self {
+        Self(shell_command(shell_cmd))
     }
 
-    pub async fn lines(&mut self) -> Result<Vec<String>> {
+    pub async fn lines(&mut self) -> std::io::Result<Vec<String>> {
         // Calling `output()` or `spawn().wait_with_output()` directly does not
         // work for Vim.
         // let output = self.0.spawn()?.wait_with_output().await?;
@@ -62,15 +61,15 @@ mod tests {
 
     #[tokio::test]
     async fn test_tokio_command() {
-        let mut tokio_cmd: TokioCommand = format!(
+        let shell_cmd = format!(
             "ls {}",
             std::env::current_dir()
                 .unwrap()
                 .into_os_string()
                 .into_string()
                 .unwrap()
-        )
-        .into();
+        );
+        let mut tokio_cmd = TokioCommand::new(shell_cmd);
         assert_eq!(
             vec!["Cargo.toml", "benches", "src"]
                 .into_iter()