+++ .vim/pack/basic/start/vim-clap/crates/maple_cli/src/stdio_server/providers/custom/dumb_jump/searcher.rs
@@ -9,22 +9,22 @@ use super::QueryInfo;
 use crate::find_usages::{
     AddressableUsage, CtagsSearcher, GtagsSearcher, QueryType, RegexSearcher, Usage, Usages,
 };
-use crate::tools::ctags::{get_language, TagsConfig};
+use crate::tools::ctags::{get_language, TagsGenerator};
 use crate::utils::ExactOrInverseTerms;
 
 /// Context for performing a search.
 #[derive(Debug, Clone, Default)]
 pub(super) struct SearchingWorker {
     pub cwd: String,
-    pub extension: String,
     pub query_info: QueryInfo,
+    pub source_file_extension: String,
 }
 
 impl SearchingWorker {
     fn ctags_search(self) -> Result<Vec<AddressableUsage>> {
-        let mut tags_config = TagsConfig::with_dir(self.cwd);
-        if let Some(language) = get_language(&self.extension) {
-            tags_config.languages(language.into());
+        let mut tags_generator = TagsGenerator::with_dir(self.cwd);
+        if let Some(language) = get_language(&self.source_file_extension) {
+            tags_generator.set_languages(language.into());
         }
 
         let QueryInfo {
@@ -33,7 +33,12 @@ impl SearchingWorker {
             filtering_terms,
         } = self.query_info;
 
-        CtagsSearcher::new(tags_config).search_usages(&keyword, &filtering_terms, query_type, true)
+        CtagsSearcher::new(tags_generator).search_usages(
+            &keyword,
+            &filtering_terms,
+            query_type,
+            true,
+        )
     }
 
     fn gtags_search(self) -> Result<Vec<AddressableUsage>> {
@@ -45,11 +50,11 @@ impl SearchingWorker {
         GtagsSearcher::new(self.cwd.into()).search_usages(
             &keyword,
             &filtering_terms,
-            &self.extension,
+            &self.source_file_extension,
         )
     }
 
-    async fn regex_search(self) -> Result<Vec<AddressableUsage>> {
+    fn regex_search(self) -> Result<Vec<AddressableUsage>> {
         let QueryInfo {
             keyword,
             filtering_terms,
@@ -57,10 +62,10 @@ impl SearchingWorker {
         } = self.query_info;
         let searcher = RegexSearcher {
             word: keyword,
-            extension: self.extension,
+            extension: self.source_file_extension,
             dir: Some(self.cwd.into()),
         };
-        searcher.search_usages(false, &filtering_terms).await
+        searcher.search_usages(false, &filtering_terms)
     }
 }
 
@@ -73,14 +78,15 @@ fn merge_all(
     let mut regex_results = regex_results;
     regex_results.retain(|r| !ctag_results.contains(r));
 
-    let mut ctag_results = ctag_results;
+    let mut results = ctag_results;
     if let Some(mut gtags_results) = maybe_gtags_results {
         regex_results.retain(|r| !gtags_results.contains(r));
-        ctag_results.append(&mut gtags_results);
+        results.append(&mut gtags_results);
     }
 
-    ctag_results.append(&mut regex_results);
-    ctag_results
+    results.append(&mut regex_results);
+
+    results
 }
 
 /// These is no best option here, each search engine has its own advantages and
@@ -118,11 +124,15 @@ impl SearchEngine {
             async move { searching_worker.ctags_search() }
         };
 
+        let regex_future = {
+            let searching_worker = searching_worker.clone();
+            async move { searching_worker.regex_search() }
+        };
+
         let addressable_usages = match self {
             SearchEngine::Ctags => searching_worker.ctags_search()?,
-            SearchEngine::Regex => searching_worker.regex_search().await?,
+            SearchEngine::Regex => searching_worker.regex_search()?,
             SearchEngine::CtagsAndRegex => {
-                let regex_future = searching_worker.regex_search();
                 let (ctags_results, regex_results) = futures::join!(ctags_future, regex_future);
 
                 merge_all(
@@ -137,7 +147,7 @@ impl SearchEngine {
                 let try_regex =
                     results.is_err() || results.as_ref().map(|r| r.is_empty()).unwrap_or(false);
                 if try_regex {
-                    searching_worker.regex_search().await?
+                    searching_worker.regex_search()?
                 } else {
                     results?
                 }
@@ -147,7 +157,6 @@ impl SearchEngine {
                     let searching_worker = searching_worker.clone();
                     async move { searching_worker.gtags_search() }
                 };
-                let regex_future = searching_worker.regex_search();
 
                 let (ctags_results, gtags_results, regex_results) =
                     futures::join!(ctags_future, gtags_future, regex_future);