+++ .vim/pack/basic/start/vim-clap/crates/types/src/source_item.rs
@@ -1,4 +1,8 @@
+use std::borrow::Cow;
+
 use pattern::{extract_file_name, extract_grep_pattern, extract_tag_name};
+
+use crate::matcher::Score;
 
 /// A tuple of match text piece (matching_text, offset_of_matching_text).
 #[derive(Debug, Clone)]
@@ -9,15 +13,6 @@ pub struct FuzzyText<'a> {
 
 impl<'a> FuzzyText<'a> {
     pub fn new(text: &'a str, matching_start: usize) -> Self {
-        Self {
-            text,
-            matching_start,
-        }
-    }
-}
-
-impl<'a> From<(&'a str, usize)> for FuzzyText<'a> {
-    fn from((text, matching_start): (&'a str, usize)) -> Self {
         Self {
             text,
             matching_start,
@@ -64,42 +59,59 @@ impl<T: AsRef<str>> From<T> for MatchScope {
     }
 }
 
-/// Text used in the matching algorithm.
-pub trait MatchingText {
-    /// Initial full text.
-    fn full_text(&self) -> &str;
+/// This trait represents the items used in the entire filter pipeline.
+pub trait ClapItem: std::fmt::Debug + Send + Sync + 'static {
+    /// Initial raw text.
+    fn raw_text(&self) -> &str;
 
-    /// Text for calculating the bonus score.
-    fn bonus_text(&self) -> &str {
-        self.full_text()
+    /// Text for the matching engine.
+    ///
+    /// Can be used to skip the leading icon, see `LineWithIcon` in `fuzzymatch-rs/src/lib.rs`.
+    fn match_text(&self) -> &str {
+        self.raw_text()
     }
 
-    /// Text for applying the fuzzy match algorithm.
+    /// Text specifically for performing the fuzzy matching, part of the entire
+    /// mathcing pipeline.
     ///
     /// The fuzzy matching process only happens when Some(_) is returned.
-    fn fuzzy_text(&self, match_scope: &MatchScope) -> Option<FuzzyText>;
+    fn fuzzy_text(&self, match_scope: MatchScope) -> Option<FuzzyText> {
+        extract_fuzzy_text(self.match_text(), match_scope)
+    }
+
+    // TODO: Each bonus can have its own range of `bonus_text`, make use of MatchScope.
+    /// Text for calculating the bonus score to tweak the initial matching score.
+    fn bonus_text(&self) -> &str {
+        self.match_text()
+    }
+
+    /// Constructs a text intended to be displayed on the screen without any decoration (truncation,
+    /// icon, etc).
+    ///
+    /// A concrete type of ClapItem can be structural to facilitate the matching process, in which
+    /// case it's necessary to make a formatted String for displaying in the end.
+    fn output_text(&self) -> Cow<'_, str> {
+        self.raw_text().into()
+    }
 }
 
-impl MatchingText for SourceItem {
-    fn full_text(&self) -> &str {
+impl ClapItem for SourceItem {
+    fn raw_text(&self) -> &str {
         &self.raw
     }
 
-    fn fuzzy_text(&self, match_scope: &MatchScope) -> Option<FuzzyText> {
-        self.get_fuzzy_text(match_scope)
+    fn fuzzy_text(&self, match_scope: MatchScope) -> Option<FuzzyText> {
+        self.fuzzy_text_or_exact_using_match_scope(match_scope)
     }
 }
 
-impl MatchingText for &str {
-    fn full_text(&self) -> &str {
-        self
-    }
-
-    fn fuzzy_text(&self, _match_scope: &MatchScope) -> Option<FuzzyText> {
-        Some(FuzzyText {
-            text: self,
-            matching_start: 0,
-        })
+// Impl [`ClapItem`] for raw String.
+//
+// In order to filter/calculate bonus for a substring instead of the whole String, a
+// custom wrapper is necessary to extract the text for matching/calculating bonus/diplaying, etc.
+impl<T: AsRef<str> + std::fmt::Debug + Send + Sync + 'static> ClapItem for T {
+    fn raw_text(&self) -> &str {
+        self.as_ref()
     }
 }
 
@@ -112,15 +124,8 @@ pub struct SourceItem {
     ///
     /// Could be initialized on creating a new [`SourceItem`].
     pub fuzzy_text: Option<(String, usize)>,
-    /// Text for displaying on a window with limited size.
-    pub display_text: Option<String>,
-}
-
-// NOTE: do not use it when you are dealing with a large number of items.
-impl From<&str> for SourceItem {
-    fn from(s: &str) -> Self {
-        String::from(s).into()
-    }
+    /// Text for displaying.
+    pub output_text: Option<String>,
 }
 
 impl From<String> for SourceItem {
@@ -128,7 +133,7 @@ impl From<String> for SourceItem {
         Self {
             raw,
             fuzzy_text: None,
-            display_text: None,
+            output_text: None,
         }
     }
 }
@@ -138,105 +143,84 @@ impl SourceItem {
     pub fn new(
         raw: String,
         fuzzy_text: Option<(String, usize)>,
-        display_text: Option<String>,
+        output_text: Option<String>,
     ) -> Self {
         Self {
             raw,
             fuzzy_text,
-            display_text,
+            output_text,
         }
     }
 
-    pub fn display_text(&self) -> &str {
-        if let Some(ref text) = self.display_text {
-            text
-        } else {
-            &self.raw
+    pub fn output_text_or_raw(&self) -> &str {
+        match self.output_text {
+            Some(ref text) => text,
+            None => &self.raw,
         }
     }
 
-    pub fn fuzzy_text_or_default(&self) -> &str {
-        if let Some((ref text, _)) = self.fuzzy_text {
-            text
-        } else {
-            &self.raw
+    pub fn fuzzy_text_or_exact_using_match_scope(
+        &self,
+        match_scope: MatchScope,
+    ) -> Option<FuzzyText> {
+        match self.fuzzy_text {
+            Some((ref text, offset)) => Some(FuzzyText::new(text, offset)),
+            None => extract_fuzzy_text(self.raw.as_str(), match_scope),
         }
     }
+}
 
-    pub fn get_fuzzy_text(&self, match_scope: &MatchScope) -> Option<FuzzyText> {
-        if let Some((ref text, offset)) = self.fuzzy_text {
-            return Some(FuzzyText::new(text, offset));
+pub fn extract_fuzzy_text(full: &str, match_scope: MatchScope) -> Option<FuzzyText> {
+    match match_scope {
+        MatchScope::Full => Some(FuzzyText::new(full, 0)),
+        MatchScope::TagName => extract_tag_name(full).map(|s| FuzzyText::new(s, 0)),
+        MatchScope::FileName => {
+            extract_file_name(full).map(|(s, offset)| FuzzyText::new(s, offset))
         }
-        let full = self.raw.as_str();
-        match match_scope {
-            MatchScope::Full => Some(FuzzyText::new(full, 0)),
-            MatchScope::TagName => extract_tag_name(full).map(|s| FuzzyText::new(s, 0)),
-            MatchScope::FileName => extract_file_name(full).map(Into::into),
-            MatchScope::GrepLine => extract_grep_pattern(full).map(Into::into),
+        MatchScope::GrepLine => {
+            extract_grep_pattern(full).map(|(s, offset)| FuzzyText::new(s, offset))
         }
     }
 }
 
 /// This struct represents the filtered result of [`SourceItem`].
 #[derive(Debug, Clone)]
-pub struct FilteredItem<T = i64> {
+pub struct MatchedItem {
     /// Tuple of (matched line text, filtering score, indices of matched elements)
-    pub source_item: SourceItem,
+    pub item: SourceItem,
     /// Filtering score.
-    pub score: T,
+    pub score: Score,
     /// Indices of matched elements.
     ///
     /// The indices may be truncated when truncating the text.
-    pub match_indices: Vec<usize>,
+    pub indices: Vec<usize>,
     /// Text for showing the final filtered result.
     ///
     /// Usually in a truncated form for fitting into the display window.
     pub display_text: Option<String>,
 }
 
-impl<I: Into<SourceItem>, T> From<(I, T, Vec<usize>)> for FilteredItem<T> {
-    fn from((item, score, match_indices): (I, T, Vec<usize>)) -> Self {
+impl MatchedItem {
+    pub fn new(item: SourceItem, score: Score, indices: Vec<usize>) -> Self {
         Self {
-            source_item: item.into(),
+            item,
             score,
-            match_indices,
+            indices,
             display_text: None,
         }
-    }
-}
-
-impl<I: Into<SourceItem>, T: Default> From<I> for FilteredItem<T> {
-    fn from(item: I) -> Self {
-        Self {
-            source_item: item.into(),
-            score: Default::default(),
-            match_indices: Default::default(),
-            display_text: None,
-        }
-    }
-}
-
-impl<T> FilteredItem<T> {
-    pub fn new<I: Into<SourceItem>>(item: I, score: T, match_indices: Vec<usize>) -> Self {
-        (item, score, match_indices).into()
-    }
-
-    /// Untruncated display text.
-    pub fn source_item_display_text(&self) -> &str {
-        self.source_item.display_text()
     }
 
     /// Maybe truncated display text.
-    pub fn display_text(&self) -> &str {
+    pub fn display_text(&self) -> Cow<str> {
         if let Some(ref text) = self.display_text {
-            text
+            text.into()
         } else {
-            self.source_item.display_text()
+            self.item.output_text()
         }
     }
 
     /// Returns the match indices shifted by `offset`.
     pub fn shifted_indices(&self, offset: usize) -> Vec<usize> {
-        self.match_indices.iter().map(|x| x + offset).collect()
+        self.indices.iter().map(|x| x + offset).collect()
     }
 }